+++ cli/src/git_util.rs
@@ -121,7 +121,7 @@ fn pinentry_get_pw(url: &str) -> Option<String> {
 #[tracing::instrument]
 fn get_ssh_keys(_username: &str) -> Vec<PathBuf> {
     let mut paths = vec![];
-    if let Some(home_dir) = dirs::home_dir() {
+    if let Ok(home_dir) = etcetera::home_dir() {
         let ssh_dir = Path::new(&home_dir).join(".ssh");
         for filename in ["id_ed25519_sk", "id_ed25519", "id_rsa"] {
             let key_path = ssh_dir.join(filename);